refactor: buildTreeData 搬進 src/lib 並改為可注入 DOM
